+++ src/fi/teknologiakerho/viipal01ja/hcode/MoveRelativeCommand.java
@@ -2,7 +2,7 @@ package fi.teknologiakerho.viipal01ja.hcode;
 
 public class MoveRelativeCommand extends PositionCommand {
 
-	public MoveRelativeCommand(int x, int y) {
+	public MoveRelativeCommand(double x, double y) {
 		super(x, y);
 	}
